--- src/main/java/com/example/domain/enums/Status.java
@@ -1,5 +1,0 @@
-package com.example.domain.enums;
-
-public enum Status {
-    ON, OFF
-}
