--- src/main/java/com/example/domain/enums/Theme.java
@@ -1,5 +1,0 @@
-package com.example.domain.enums;
-
-public enum Theme {
-    ON, OFF
-}
