+++ src/main/java/com/example/spot/domain/enums/Gender.java
@@ -1,4 +1,4 @@
-package com.example.domain.enums;
+package com.example.spot.domain.enums;
 
 public enum Gender {
     MALE, FEMALE, UNKNOWN
