+++ src/main/java/com/example/spot/domain/enums/Status.java
@@ -1,0 +1,5 @@
+package com.example.spot.domain.enums;
+
+public enum Status {
+    ON, OFF
+}
