+++ src/main/java/com/example/spot/domain/enums/StudyState.java
@@ -1,4 +1,4 @@
-package com.example.domain.enums;
+package com.example.spot.domain.enums;
 
 public enum StudyState {
   RECRUITING, COMPLETED, BEFORE
