+++ src/main/java/com/example/spot/domain/enums/Theme.java
@@ -1,0 +1,5 @@
+package com.example.spot.domain.enums;
+
+public enum Theme {
+    ON, OFF
+}
