+++ src/main/java/com/example/spot/domain/study/Option.java
@@ -1,16 +1,16 @@
-package com.example.domain.study;
+package com.example.spot.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.mapping.StudyLikedComment;
-import com.example.domain.study.StudyPost;
+import com.example.spot.domain.mapping.mapping.MemberVote;
 import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -27,24 +27,18 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyPostComment extends BaseEntity {
+@Table(name = "options")
+public class Option extends BaseEntity {
 
-    @Id @GeneratedValue
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    private StudyPost studyPost;
+    private Vote vote;
 
-    // @Column(nullable = false)
-    // private Member member;
-
-    @Column(nullable = false)
     private String content;
 
-    @Column(nullable = false)
-    private Integer likeCount;
-
-    @OneToMany(mappedBy = "studyPostComment", cascade = CascadeType.ALL)
-    private List<StudyLikedComment> likedComments;
+    @OneToMany(mappedBy = "option", cascade = CascadeType.ALL)
+    private List<MemberVote> memberVotes;
 
 }
