+++ src/main/java/com/example/spot/domain/study/Schedule.java
@@ -1,13 +1,14 @@
-package com.example.domain.mapping;
+package com.example.spot.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.study.StudyPostComment;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -23,14 +24,24 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyLikedComment extends BaseEntity {
+public class Schedule extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    private StudyPostComment studyPostComment;
+    private Study study;
 
-    //private Member member;
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String location;
+
+    @Column(nullable = false)
+    private LocalDateTime staredAt;
+
+    @Column(nullable = false)
+    private LocalDateTime finishedAt;
 
 }
