+++ src/main/java/com/example/spot/domain/study/StudyPostComment.java
@@ -1,13 +1,16 @@
-package com.example.domain.mapping;
+package com.example.spot.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.study.StudyPost;
+import com.example.spot.domain.mapping.mapping.StudyLikedComment;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -23,15 +26,24 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class StudyLikedPost extends BaseEntity {
+public class StudyPostComment extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private StudyPost studyPost;
 
-    //private Member member;
+    // @Column(nullable = false)
+    // private Member member;
+
+    @Column(nullable = false)
+    private String content;
+
+    @Column(nullable = false)
+    private Integer likeCount;
+
+    @OneToMany(mappedBy = "studyPostComment", cascade = CascadeType.ALL)
+    private List<StudyLikedComment> likedComments;
 
 }
