+++ src/main/java/com/example/spot/domain/study/Vote.java
@@ -1,13 +1,16 @@
-package com.example.domain.mapping;
+package com.example.spot.domain.study;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.study.Option;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.time.LocalDateTime;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -23,13 +26,24 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class MemberVote extends BaseEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+public class Vote extends BaseEntity {
+    @Id @GeneratedValue
     private long id;
 
-   // private Member member
-
     @ManyToOne(fetch = FetchType.LAZY)
-    private Option option;
+    private Study study;
+
+    //private Member member;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private boolean isMultipleChoice;
+
+    @Column(nullable = false)
+    private LocalDateTime finishedAt;
+
+    @OneToMany(mappedBy = "vote", cascade = CascadeType.ALL)
+    private List<Option> options;
 }
